Adapting Frame.extractor_chain method to work with new chunk-based processing

+++ zounds/analyze2/__init__.py
@@ -1,2 +1,2 @@
 # KLUDGE: This needs to be an environment variable
-chunksize = 5.0 * 44100
+chunksize = int(5.0 * 44100)
